fix: circuitbreaker boltdb load int fail (#903)
* feat: upgrade specification to 1.1.0 stable release

* feat：适配V1规则的分组名，V1规则分组名=V2规则名.分组名

* fix: boltdb load int value should be int64

+++ store/boltdb/circuitbreaker_rule.go
@@ -256,7 +256,7 @@ func (c *circuitBreakerStore) GetCircuitBreakerRules(
 					var inLevel = false
 					for _, level := range levels {
 						levelInt, _ := strconv.Atoi(level)
-						if levelInt == fieldValue.(int) {
+						if int64(levelInt) == fieldValue.(int64) {
 							inLevel = true
 							break
 						}
